IO-1973 Adding in seperate Simulate files under test utils
