+++ src/TestUtils/index.ts
@@ -1,4 +1,11 @@
 export * from "./EnzymeHelper";
 export * from "./FakeComponent";
 export * from "./SeedHelper";
+export * from "./SimulateCheckboxInput";
+export * from "./SimulateFormInput";
+export * from "./SimulateOnBlur";
+export * from "./SimulateOnFocus";
+export * from "./SimulateRadioInput";
+export * from "./SimulateSelection";
+export * from "./SimulateTextInput";
 export * from "./UIEventSimulation";
